Initial Commit
Some of the way this program is done is simply because that's the way
the teach wanted it done. Personally I would be the load into the
constructor. Classes shouldn't have an "initialize" function, that's
what a constructor is for. The int datastructures were provided by the
teacher. I did not code any part of them.